Struct multistr::vec::StringVec

pub struct StringVec {
    buffer: String,
    split: Vec<usize>,
}

Vec of immutable strings stored on the heap in the same buffer.

Slicing ranges of the vector yields the strings in the range concatenated together.

Fields
buffer: String split: Vec<usize>
Methods

impl StringVec

fn new() -> StringVec

Creates an empty StringVec.

fn with_capacities(num: usize, bytes: usize) -> StringVec

Creates an empty StringVec with the given capacities.

The StringVec will be able to hold exactly count strings totallying up to bytes in length without reallocating. If count and bytes are zero, the vector will not allocate.

fn num_capacity(&self) -> usize

Returns the number of strings this vector can hold without reallocating.

fn byte_capacity(&self) -> usize

Returns the number of bytes this vector can hold without reallocating.

fn reserve(&mut self, additional: usize, bytes: usize)

Reserves capacity for at least additional more strings totalling to bytes more bytes.

fn reserve_exact(&mut self, additional: usize, bytes: usize)

Similar to reserve, calling reserve_exact on the inner String and Vec.

fn shrink_to_fit(&mut self)

See: Vec::shrink_to_fit.

fn truncate(&mut self, len: usize)

Shortens the vector, keeping the first len strings and dropping the rest.

fn append(&mut self, other: &mut StringVec)

Moves all of the elements of other into self, leaving other empty.

fn len(&self) -> usize

Returns the number of strings in the vector.

fn is_empty(&self) -> bool

Returns true iff the vector contains no elements.

fn split_off(&mut self, at: usize) -> StringVec

Splits the collection into two at the given index.

fn clear(&mut self)

Clears the vector, removing all strings.

fn push<S: AsRef<str>>(&mut self, s: S)

Adds a string to the end of the vec.

fn pop(&mut self) -> bool

Removes a string from the end of the vec and discards it.

fn pop_off(&mut self) -> Option<String>

Removes a string from the end of the vec and allocates it onto a new buffer.

fn iter(&self) -> Iter

Returns an iterator over the strings in the vector.

fn get_start_idx(&self, index: usize) -> usize

fn get_end_idx(&self, index: usize) -> usize